Add support for scaling recordings (#11)
This adds support for scaling the resolution of recordings, but it is slooooow.  It seems it's actually the encoding process that is slow though, not the actual scaling.  It also results in much larger animated PNGs, which I dislike.

Because of these issues, I'm just defaulting the scale to 1 for now, but this can be changed in the `SCALE_FACTOR` constant in `recorder-ui.ts`.